+++ kuStudy/kuStudy Today Extension/LibraryTableViewCell.swift
@@ -10,25 +10,30 @@ import UIKit
 import kuStudyKit
 
 class LibraryTableViewCell: UITableViewCell {
-    @IBOutlet weak var libraryImageView: RingImageView!
+    @IBOutlet weak var thumbnailImageView: UIImageView!
+    @IBOutlet weak var percentageView: CircularProgressView!
     @IBOutlet weak var nameLabel: UILabel!
     @IBOutlet weak var availableLabel: UILabel!
     
     // MARK: Setup
     override func awakeFromNib() {
         super.awakeFromNib()
-        libraryImageView.layer.cornerRadius = libraryImageView.bounds.width / 2
+        thumbnailImageView.layer.cornerRadius = thumbnailImageView.bounds.width / 2
+        percentageView.progressBackgroundColor = UIColor(hue: 0, saturation: 0, brightness: 0.95, alpha: 1)
     }
     
     // MARK: Populate
     func populate(libraryData: LibraryData) {
-        guard let libraryId = libraryData.libraryId else { return }
+        guard let libraryId = libraryData.libraryId,
+            availableSeats = libraryData.availableSeats
+            else { return }
+        
         let libraryType = LibraryType(rawValue: libraryId)
         nameLabel.text = libraryType?.name
-        availableLabel.text = libraryData.availableSeats?.readableFormat
+        availableLabel.text = availableSeats.readableFormat + " " + "Available"
         
-        libraryImageView.ringColor = libraryData.availablePercentageColor
-        libraryImageView.rating = libraryData.availablePercentage
-        libraryImageView.image = libraryData.thumbnail
+        thumbnailImageView.image = libraryData.thumbnail
+        percentageView.progress = libraryData.availablePercentage
+        percentageView.progressColor = libraryData.availablePercentageColor
     }
 }