+++ kuStudy/kuStudy/extension/UIColor.swift
@@ -1,0 +1,15 @@
+//
+//  UIColor.swift
+//  kuStudy
+//
+//  Created by BumMo Koo on 2016. 8. 6..
+//  Copyright © 2016년 gbmKSquare. All rights reserved.
+//
+
+import UIKit
+
+extension UIColor {
+    static var themeColor: UIColor {
+        return UIColor(hue: 359/359, saturation: 0.78, brightness: 0.83, alpha: 1)
+    }
+}