+++ kuStudy/kuStudy/extension/UINavigationController.swift
@@ -1,8 +1,8 @@
 //
-//  Extensions.swift
+//  UINavigationController.swift
 //  kuStudy
 //
-//  Created by 구범모 on 2016. 1. 25..
+//  Created by BumMo Koo on 2016. 8. 6..
 //  Copyright © 2016년 gbmKSquare. All rights reserved.
 //
 
@@ -15,11 +15,5 @@ extension UINavigationController {
         navigationBar.shadowImage = UIImage()
         navigationBar.backgroundColor = UIColor.clearColor()
         view.backgroundColor = UIColor.clearColor()
-    }
-}
-
-extension UIColor {
-    class func navigationColor() -> UIColor {
-        return UIColor(hue: 359/359, saturation: 0.78, brightness: 0.83, alpha: 1)
     }
 }